requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
fpdf
gspread
google-auth
requests
gunicorn
